Data Pipelines/ReallyBigJobData_Pipeline/requirements.txt
psycopg2-binary
pandas
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
python-dateutil
scikit-learn
joblib
pokemontcgsdk
requests
warcio
kafka-python
pyspark
apache-airflow